Port Twelf to SML Standard Basis 2002.
Keep backwards-compatibility with SML/NJ 110.0.7 (which uses Basis
'97) and SML/NJ 110.4x (which mix Bases '97 and '02) through a Compat
structure that reflects the Basis '02 API, but uses CM conditional
compilation to incorporate compatibility code when necessary.

Consequently remove the '-basis 1997' switch in the MLton makefile, as
the lastest MLton release defaults to Basis '02.


git-svn-id: https://svn.concert.cs.cmu.edu/twelf/trunk@994 92263adb-174b-0410-95c5-c49bc0650b85