[CI] Add Github Actions workflow to run `regression.sh`
A docker image built from https://github.com/jcreedcmu/docker-mlton is
available as
[jcreed/mlton-twelf-ci:latest](https://hub.docker.com/repository/docker/jcreed/mlton-twelf-ci/general)
on Docker Hub. It contains the MLton release binaries, and a script
placed in `/usr/bin/svnversion` that actually just asks for the
current git revision since we live in the present time and no longer
use stone tools. This is enough for `TEST/regression.sh` to
successfully run.

In passing, tweaked the bash syntax of that regression script, and
reduced chatter in the warnings generated by MLton compilation of
twelf, presumably due to the evolution of the semantics of
`default-ann 'nonexhaustiveMatch ignore'` in the intervening years.