Add static site generator for future twelf.org (#11)
This PR moves most of the content of the MediaWiki installation of twelf.org into the repository ask a collection of Twelf files (/wiki/pages) that are converted by a script (wiki/elf-to-mdx.mjs and wiki/elf-watcher.mjs) to MDX files that can be built with Starlight, a documentation-focused extension to the Astro static site generator. This is an entirely self-contained addition with the exception of the regression tests, which now test the actual contents of the wiki rather than a manually-pulled snapshot.