Produce static site on GitHub and upload as artifact (#13)
Taking advantage of #12 means that we need to have Twelf around while the wiki is being statically site generated, but it seems like the build infrastructure for static sites on render.com doesn't have the easy ability to install extra dependencies, as would be necessary to compile Twelf.

One can imagine several possible ways forward, the most tempting being using the inherent portability of WASM and the fact that we have a WASM twelf running around. But we also have Twelf compiling successfully with github workflows... so I'm going to take the approach of just moving the static site generation step to GitHub actions, and have the render.com process doing nothing more than polling GitHub for the right artifact.