Temp workaround for Astro performance (#25)
Indexed HOAS nat bijection and (to a lesser degree) HOAS nat bijection
choke out the MDX parser underlying Astro. Clearly there's some
quadratic behavior in the multi-language parsing stack that hasn't been
dealt with, so the 500kb elf file that becomes a 16mb MDX file takes an
hour to load. Clearly this is absurd, and I need to send and upstream
bug report, but in the meantime this is a workaround.

Also an attempt to use a GitHub action to trigger the render.com
workflow by API call, because previously the render.com had just been
idling and using build minutes waiting for GitHub to finish after
deploy, and when that became a 60 minute process I turned off
auto-deploy.