Use the filesystem to cache Twelf tag contexts (#29)
This reduces intermediate and final asset sizes substantially, and
allows the indexed-hoas-nat-bijection files to be reintroduced despite
performance issues with very large mdx files.

The fact that `reduceTwelfAccum` is async is a footgun - the generated
mdx was just weirdly wrong until I awaited all calls to that function,
so now I know that I do not like async functions that don't return
meaningful values. Aside from that I'm reasonably happy with this
design.