Wiki fixes (#51)
I've made three scripted changes to the wiki here. The first change
converts the older `&lt;math&gt;` equations into `<Math>` and
`<DisplayMath>` as appropriate (display if proceeded by two newlines).
The second change replaces the symbol templates like `\{\{darrow\}\}`
with an appropriate unicode character. And the third change is to
replace `&gt;` in literals with `>`.

There are still issues like table tags in `admissibility-of-cut.elf`
that may need manual attention, but hopefully this will save a lot of
work in updating the wiki.